Tested/Implemented Plane.

+++ src/main/java/uk/bs338/codeclan/javaAir/PlaneType.java
@@ -3,19 +3,19 @@ package uk.bs338.codeclan.javaAir;
 public enum PlaneType {
     BOEING747(5, 20);
 
-    private final int capacity;
-    private final int totalWeight;
+    private final double capacity;
+    private final double totalWeight;
 
     PlaneType(int capacity, int totalWeight) {
         this.capacity = capacity;
         this.totalWeight = totalWeight;
     }
 
-    public int getCapacity() {
+    public double getCapacity() {
         return capacity;
     }
 
-    public int getTotalWeight() {
+    public double getTotalWeight() {
         return totalWeight;
     }
 }
